fashion-mcp/services/fullbody-validator/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pillow==11.3.0
numpy==1.26.4
mediapipe==0.10.18
pytest==8.4.1
httpx==0.28.1
